pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "correios-reverso"
version = "0.2.0"
description = "Automacao da plataforma Pre-Postagem dos Correios (REST API + MCP Server)"
readme = "README.md"
keywords = ["correios", "logistica", "reversa", "pre-postagem", "etiqueta", "cep", "brasil"]
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Nicolas Pastorello", email = "user@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/opastorello/correios-reverso"
Repository = "https://github.com/opastorello/correios-reverso"
Issues = "https://github.com/opastorello/correios-reverso/issues"
Documentation = "https://github.com/opastorello/correios-reverso#readme"
Changelog = "https://github.com/opastorello/correios-reverso/releases"

[project.optional-dependencies]
api = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "responses>=0.25.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: testes que fazem chamadas reais",
    "integration: testes de integração end-to-end",
]
